LayoutTests/imported/w3c/web-platform-tests/fetch/api/abort/general-sharedworker-expected.txt

CONSOLE MESSAGE: line 11: ReferenceError: Can't find variable: SharedWorker

FAIL General fetch abort tests - shared worker ReferenceError: Can't find variable: SharedWorker
